Move action_view to openspiel core.
