+++ open_spiel/simultaneous_move_game.cc
@@ -19,9 +19,9 @@
 #include <vector>
 
 #include "open_spiel/abseil-cpp/absl/strings/str_cat.h"
+#include "open_spiel/action_view.h"
 #include "open_spiel/spiel.h"
 #include "open_spiel/spiel_utils.h"
-#include "open_spiel/utils/action_view.h"
 
 namespace open_spiel {
